@@ -1,6 +1,10 @@
-import React from "react";
-import PortfolioContainer from "./components/PortfolioContainer";
+import logo from './assets/LOGO.png'
+import PortfolioContainer from './components/PortfolioContainer';
 
-const App = () => <PortfolioContainer />;
+function App() {
+  return (
+    <PortfolioContainer />
+  );
+}
 
 export default App;
