@@ -1,23 +1,23 @@
-import React, { useState } from 'react';
-import Header from './Header';
-import Footer from './Footer';
-import Contact from './pages/Contact';
-import About from './pages/About';
-import Projects from './pages/Projects';
-import Resume from './pages/Resume';
+import React, { useState } from "react";
+import Header from "./Header";
+import Footer from "./Footer";
+import Contact from "./pages/Contact";
+import About from "./pages/About";
+import Projects from "./pages/Projects";
+import Resume from "./pages/Resume";
 
 export default function PortfolioContainer() {
-  const [currentPage, setCurrentPage] = useState('About');
+  const [currentPage, setCurrentPage] = useState("About");
 
   // This method is checking to see what the value of `currentPage` is. Depending on the value of currentPage, we return the corresponding component to render.
   const renderPage = () => {
-    if (currentPage === 'Projects') {
+    if (currentPage === "Projects") {
       return <Projects />;
     }
-    if (currentPage === 'Resume') {
+    if (currentPage === "Resume") {
       return <Resume />;
     }
-    if (currentPage === 'Contact') {
+    if (currentPage === "Contact") {
       return <Contact />;
     }
     return <About />;
@@ -26,12 +26,13 @@ export default function PortfolioContainer() {
   const handlePageChange = (page) => setCurrentPage(page);
 
   return (
-    <div className='bg-[#191825] relative h-screen'>
+    <div className="bg-d-pur relative h-screen">
       {/* We are passing the currentPage from state and the function to update it */}
       <Header currentPage={currentPage} handlePageChange={handlePageChange} />
       {/* Here we are calling the renderPage method which will return a component  */}
       {renderPage()}
       <Footer />
     </div>
+    
   );
 }
