@@ -2,16 +2,10 @@ import React from 'react';
 
 export default function About() {
   return (
-    <div className=''>
-      <h1>About Page</h1>
-      <p>
-        Nunc pharetra finibus est at efficitur. Praesent sed congue diam.
-        Integer gravida dui mauris, ut interdum nunc egestas sed. Aenean sed
-        mollis diam. Nunc aliquet risus ac finibus porta. Nam quis arcu non
-        lectus tincidunt fermentum. Suspendisse aliquet orci porta quam semper
-        imperdiet. Praesent euismod mi justo, faucibus scelerisque risus cursus
-        in. Sed rhoncus mollis diam, sit amet facilisis lectus blandit at.
-      </p>
+    <div className='bg-d-pur font-standard p-4 pl-10 pr-10'>
+      <div className='text-9xl'><h1>Hey,</h1><h1> My name is </h1><h1 className='italic '>Sabrina</h1></div>
+      <div className='text-6xl pt-10'><h1> Full Stack Web Developer</h1></div>
+       
     </div>
   );
 }
